refactor: pull components into different files
