+++ src/app/page.tsx
@@ -2,358 +2,16 @@
 
 import { useEffect, useRef, useState } from "react"
 
-import { cn } from "@/lib/util/cn"
+import { RunStats } from "@/components/run-stats"
+import { Sentence } from "@/components/sentence"
 import { pusher, type Channel } from "@/lib/util/pusher/client"
-
-const words = {
-  easy: [
-    "the",
-    "and",
-    "you",
-    "that",
-    "was",
-    "for",
-    "are",
-    "with",
-    "his",
-    "they",
-    "have",
-    "this",
-    "from",
-    "one",
-    "had",
-    "word",
-    "but",
-    "not",
-    "what",
-    "all",
-    "were",
-    "we",
-    "when",
-    "your",
-    "can",
-    "said",
-    "there",
-    "use",
-    "an",
-    "each",
-    "which",
-    "she",
-    "do",
-    "how",
-    "their",
-    "if",
-    "will",
-    "up",
-    "other",
-    "about",
-    "out",
-    "many",
-    "then",
-    "them",
-    "these",
-    "so",
-    "some",
-    "her",
-    "would",
-    "make",
-    "like",
-    "him",
-    "into",
-    "time",
-    "has",
-    "look",
-    "two",
-    "more",
-    "write",
-    "see",
-    "number",
-    "way",
-    "could",
-    "people",
-    "than",
-    "first",
-    "water",
-    "been",
-    "call",
-    "who",
-    "oil",
-    "its",
-    "now",
-    "find",
-    "long",
-    "down",
-    "day",
-    "did",
-    "get",
-    "come",
-    "made",
-    "may",
-    "part",
-    "over",
-    "new",
-    "sound",
-    "take",
-    "only",
-    "little",
-    "work",
-    "know",
-    "place",
-    "year",
-    "live",
-    "me",
-    "back",
-    "give",
-    "most",
-    "very",
-  ],
-
-  medium: [
-    "meadow",
-    "tumour",
-    "complain",
-    "describe",
-    "constant",
-    "terrace",
-    "representative",
-    "similar",
-    "introduction",
-    "systematic",
-    "reserve",
-    "useful",
-    "orientation",
-    "loyalty",
-    "happen",
-    "generate",
-    "ditch",
-    "uniform",
-    "bench",
-    "rocket",
-    "method",
-    "congress",
-    "drama",
-    "coincide",
-    "industry",
-    "communication",
-    "committee",
-    "sculpture",
-    "animal",
-    "workshop",
-    "hallway",
-    "attack",
-    "autonomy",
-    "attitude",
-    "appetite",
-    "legislation",
-    "flash",
-    "persist",
-    "skilled",
-    "majority",
-    "border",
-    "support",
-    "blank",
-    "shell",
-    "remind",
-    "retirement",
-    "tower",
-    "thick",
-    "proper",
-    "expression",
-    "although",
-    "beginning",
-    "children",
-    "develop",
-    "everybody",
-    "friendship",
-    "happiness",
-    "important",
-    "knowledge",
-    "literature",
-    "maintenance",
-    "necessary",
-    "operation",
-    "particular",
-    "questioning",
-    "realization",
-    "satisfaction",
-    "themselves",
-    "understand",
-    "vocabulary",
-    "wonderful",
-    "beautiful",
-    "character",
-    "delicious",
-    "elephant",
-    "furniture",
-    "gratitude",
-    "happening",
-    "influence",
-    "journey",
-    "kitchen",
-    "landscape",
-    "measurement",
-    "opportunity",
-    "photography",
-    "quantity",
-    "reflection",
-    "sincerely",
-    "technology",
-    "university",
-    "valuable",
-    "wondering",
-    "yellow",
-    "abundance",
-    "boundary",
-    "celebration",
-    "difference",
-    "efficiency",
-    "flourish",
-    "generation",
-    "harmony",
-    "incredible",
-    "judgment",
-    "knowledgeable",
-    "leadership",
-    "mechanism",
-    "narrative",
-    "occurrence",
-    "parallel",
-    "quintessential",
-    "reliability",
-    "transformation",
-    "unbelievable",
-    "vibrant",
-    "wholesome",
-    "xylophone",
-    "yielding",
-    "zealous",
-    "aberration",
-    "blissful",
-    "capitulate",
-    "dexterity",
-    "effervescent",
-    "flamboyant",
-    "garrulous",
-    "iconoclast",
-    "juxtapose",
-    "kaleidoscope",
-    "labyrinthine",
-    "magnanimous",
-    "nebulous",
-    "obfuscate",
-    "peculiar",
-    "quixotic",
-    "recalcitrant",
-    "serendipity",
-    "trepidation",
-    "ubiquitous",
-    "verisimilitude",
-    "whimsical",
-    "xenophobe",
-    "yesteryear",
-    "zeppelin",
-  ],
-
-  hard: [
-    "exaggeration",
-    "quizzically",
-    "labyrinthine",
-    "haphazardly",
-    "bewilderment",
-    "equivocal",
-    "discombobulated",
-    "inscrutable",
-    "juxtaposition",
-    "perfidious",
-    "serendipity",
-    "cacophony",
-    "effervescent",
-    "phosphorescent",
-    "onomatopoeia",
-    "incandescent",
-    "paradoxical",
-    "ephemeral",
-    "lackadaisical",
-    "ubiquitous",
-    "procrastination",
-    "pusillanimous",
-    "idiosyncrasy",
-    "sesquipedalian",
-    "superfluous",
-    "facetious",
-    "garrulous",
-    "mellifluous",
-    "recalcitrant",
-    "antithesis",
-    "grandiloquent",
-    "disquisition",
-    "perspicacious",
-    "ubiquity",
-    "pellucid",
-    "aberration",
-    "capitulate",
-    "circumlocution",
-    "disparate",
-    "ebullient",
-    "fastidious",
-    "gregarious",
-    "iconoclast",
-    "jaundiced",
-    "kaleidoscope",
-    "lugubrious",
-    "magnanimous",
-    "nonchalant",
-    "obfuscate",
-    "quintessential",
-    "supercilious",
-    "tenebrous",
-    "umbrage",
-    "verisimilitude",
-    "welter",
-    "xenophobe",
-    "yammering",
-    "zeppelin",
-    "ambidextrous",
-    "belligerent",
-    "cogent",
-    "defenestration",
-    "quixotic",
-  ],
-}
-
-function generateRandomSentence({
-  difficulty,
-  length = 10,
-}: {
-  difficulty: "easy" | "medium" | "hard"
-  length?: number
-}) {
-  const available_words = words[difficulty]
-  let sentence: string[] = []
-
-  for (let i = 0; i < length; i++) {
-    // Pick a random word
-    const r = Math.floor(Math.random() * available_words.length)
-    const newWord = available_words[r]
-
-    // Compare the word to the last word, reset `i` if necessary
-    const lastWord = sentence.at(-1)
-    if (newWord === lastWord) {
-      i -= 1
-      continue
-    }
-
-    sentence.push(available_words[r])
-  }
-
-  return sentence.join(" ")
-}
+import { generateRandomSentence } from "@/lib/util/sentences/generateRandomSentence"
 
 export default function Home() {
   const [sentence, setSentence] = useState("")
   const [userInput, setUserInput] = useState("")
   const [startTime, setStartTime] = useState<number | null>(null)
-  const [stats, setStats] = useState<Stats | null>(null)
+  const [stats, setStats] = useState<RunInfo | null>(null)
   const gameChannel = useRef<Channel | null>(null)
 
   const generateSentence = () => {
@@ -419,103 +77,8 @@ export default function Home() {
 
   return (
     <main className="grid h-[100svh] place-content-center backdrop-blur-xl">
-      {/* <Statistics stats={stats ?? { duration: 0, numChars: 0 }} />
-      <SentenceView sentence={sentence} userInput={userInput} /> */}
-
-      <button
-        onClick={() => {
-          gameChannel.current!.trigger("client-test", {})
-        }}
-      >
-        hit
-      </button>
+      <RunStats runInfo={stats ?? { duration: 0, numChars: 0 }} />
+      <Sentence sentence={sentence} userInput={userInput} />
     </main>
-  )
-}
-
-function SentenceView({
-  sentence,
-  userInput,
-}: {
-  sentence: string
-  userInput: string
-}) {
-  const sentenceChars = sentence.split("")
-  const userInputChars = userInput.split("")
-  const currentIndex = userInput.length
-
-  return (
-    <div className="flex flex-row gap-x-1 text-5xl">
-      {sentenceChars.map((expected, i) => (
-        <Character
-          key={expected + i}
-          expected={expected}
-          received={userInputChars[i] ?? null}
-          current={i === currentIndex}
-        />
-      ))}
-    </div>
-  )
-}
-
-function Character({
-  expected,
-  received,
-  current = false,
-}: {
-  expected: string
-  received: string | null
-  current?: boolean
-}) {
-  const classNames = current
-    ? "text-black bg-yellow-300 animate-pulse rounded px-1"
-    : !received
-      ? "text-white/40"
-      : expected === received
-        ? "text-emerald-500/70"
-        : "text-rose-500/70"
-
-  if (expected === " ")
-    return (
-      <span
-        className={cn(
-          "h-16 w-[calc(1ch)] -translate-y-2 rounded",
-          current && classNames + " w-[calc(1ch+0.5rem)]",
-          received && received !== expected && "bg-rose-500/80",
-          "px-0",
-        )}
-      />
-    )
-  else
-    return (
-      <div className="relative">
-        <span className={classNames}>{expected}</span>
-
-        {received ? (
-          <span
-            className={cn(
-              "absolute right-1/2 top-1/2 -translate-y-1/2 translate-x-1/2 blur-lg",
-              classNames,
-            )}
-          >
-            {expected}
-          </span>
-        ) : null}
-      </div>
-    )
-}
-
-type Stats = {
-  duration: number
-  numChars: number
-}
-function Statistics({ stats: { duration, numChars } }: { stats: Stats }) {
-  return (
-    <div className="mb-4 flex flex-row gap-8 text-3xl">
-      <span>{(duration / 1000).toFixed(2)}s</span>
-      <span>{(numChars / 5 / (duration / 60000)).toFixed(2)}wpm</span>
-
-      <span>{numChars} chars</span>
-    </div>
   )
 }
