+++ src/lib/util/sentences/generateRandomSentence.ts
@@ -1,0 +1,343 @@
+const words = {
+  easy: [
+    "the",
+    "and",
+    "you",
+    "that",
+    "was",
+    "for",
+    "are",
+    "with",
+    "his",
+    "they",
+    "have",
+    "this",
+    "from",
+    "one",
+    "had",
+    "word",
+    "but",
+    "not",
+    "what",
+    "all",
+    "were",
+    "we",
+    "when",
+    "your",
+    "can",
+    "said",
+    "there",
+    "use",
+    "an",
+    "each",
+    "which",
+    "she",
+    "do",
+    "how",
+    "their",
+    "if",
+    "will",
+    "up",
+    "other",
+    "about",
+    "out",
+    "many",
+    "then",
+    "them",
+    "these",
+    "so",
+    "some",
+    "her",
+    "would",
+    "make",
+    "like",
+    "him",
+    "into",
+    "time",
+    "has",
+    "look",
+    "two",
+    "more",
+    "write",
+    "see",
+    "number",
+    "way",
+    "could",
+    "people",
+    "than",
+    "first",
+    "water",
+    "been",
+    "call",
+    "who",
+    "oil",
+    "its",
+    "now",
+    "find",
+    "long",
+    "down",
+    "day",
+    "did",
+    "get",
+    "come",
+    "made",
+    "may",
+    "part",
+    "over",
+    "new",
+    "sound",
+    "take",
+    "only",
+    "little",
+    "work",
+    "know",
+    "place",
+    "year",
+    "live",
+    "me",
+    "back",
+    "give",
+    "most",
+    "very",
+  ],
+
+  medium: [
+    "meadow",
+    "tumour",
+    "complain",
+    "describe",
+    "constant",
+    "terrace",
+    "representative",
+    "similar",
+    "introduction",
+    "systematic",
+    "reserve",
+    "useful",
+    "orientation",
+    "loyalty",
+    "happen",
+    "generate",
+    "ditch",
+    "uniform",
+    "bench",
+    "rocket",
+    "method",
+    "congress",
+    "drama",
+    "coincide",
+    "industry",
+    "communication",
+    "committee",
+    "sculpture",
+    "animal",
+    "workshop",
+    "hallway",
+    "attack",
+    "autonomy",
+    "attitude",
+    "appetite",
+    "legislation",
+    "flash",
+    "persist",
+    "skilled",
+    "majority",
+    "border",
+    "support",
+    "blank",
+    "shell",
+    "remind",
+    "retirement",
+    "tower",
+    "thick",
+    "proper",
+    "expression",
+    "although",
+    "beginning",
+    "children",
+    "develop",
+    "everybody",
+    "friendship",
+    "happiness",
+    "important",
+    "knowledge",
+    "literature",
+    "maintenance",
+    "necessary",
+    "operation",
+    "particular",
+    "questioning",
+    "realization",
+    "satisfaction",
+    "themselves",
+    "understand",
+    "vocabulary",
+    "wonderful",
+    "beautiful",
+    "character",
+    "delicious",
+    "elephant",
+    "furniture",
+    "gratitude",
+    "happening",
+    "influence",
+    "journey",
+    "kitchen",
+    "landscape",
+    "measurement",
+    "opportunity",
+    "photography",
+    "quantity",
+    "reflection",
+    "sincerely",
+    "technology",
+    "university",
+    "valuable",
+    "wondering",
+    "yellow",
+    "abundance",
+    "boundary",
+    "celebration",
+    "difference",
+    "efficiency",
+    "flourish",
+    "generation",
+    "harmony",
+    "incredible",
+    "judgment",
+    "knowledgeable",
+    "leadership",
+    "mechanism",
+    "narrative",
+    "occurrence",
+    "parallel",
+    "quintessential",
+    "reliability",
+    "transformation",
+    "unbelievable",
+    "vibrant",
+    "wholesome",
+    "xylophone",
+    "yielding",
+    "zealous",
+    "aberration",
+    "blissful",
+    "capitulate",
+    "dexterity",
+    "effervescent",
+    "flamboyant",
+    "garrulous",
+    "iconoclast",
+    "juxtapose",
+    "kaleidoscope",
+    "labyrinthine",
+    "magnanimous",
+    "nebulous",
+    "obfuscate",
+    "peculiar",
+    "quixotic",
+    "recalcitrant",
+    "serendipity",
+    "trepidation",
+    "ubiquitous",
+    "verisimilitude",
+    "whimsical",
+    "xenophobe",
+    "yesteryear",
+    "zeppelin",
+  ],
+
+  hard: [
+    "exaggeration",
+    "quizzically",
+    "labyrinthine",
+    "haphazardly",
+    "bewilderment",
+    "equivocal",
+    "discombobulated",
+    "inscrutable",
+    "juxtaposition",
+    "perfidious",
+    "serendipity",
+    "cacophony",
+    "effervescent",
+    "phosphorescent",
+    "onomatopoeia",
+    "incandescent",
+    "paradoxical",
+    "ephemeral",
+    "lackadaisical",
+    "ubiquitous",
+    "procrastination",
+    "pusillanimous",
+    "idiosyncrasy",
+    "sesquipedalian",
+    "superfluous",
+    "facetious",
+    "garrulous",
+    "mellifluous",
+    "recalcitrant",
+    "antithesis",
+    "grandiloquent",
+    "disquisition",
+    "perspicacious",
+    "ubiquity",
+    "pellucid",
+    "aberration",
+    "capitulate",
+    "circumlocution",
+    "disparate",
+    "ebullient",
+    "fastidious",
+    "gregarious",
+    "iconoclast",
+    "jaundiced",
+    "kaleidoscope",
+    "lugubrious",
+    "magnanimous",
+    "nonchalant",
+    "obfuscate",
+    "quintessential",
+    "supercilious",
+    "tenebrous",
+    "umbrage",
+    "verisimilitude",
+    "welter",
+    "xenophobe",
+    "yammering",
+    "zeppelin",
+    "ambidextrous",
+    "belligerent",
+    "cogent",
+    "defenestration",
+    "quixotic",
+  ],
+}
+
+export function generateRandomSentence({
+  difficulty,
+  length = 10,
+}: {
+  difficulty: "easy" | "medium" | "hard"
+  length?: number
+}) {
+  const available_words = words[difficulty]
+  let sentence: string[] = []
+
+  for (let i = 0; i < length; i++) {
+    // Pick a random word
+    const r = Math.floor(Math.random() * available_words.length)
+    const newWord = available_words[r]
+
+    // Compare the word to the last word, reset `i` if necessary
+    const lastWord = sentence.at(-1)
+    if (newWord === lastWord) {
+      i -= 1
+      continue
+    }
+
+    sentence.push(available_words[r])
+  }
+
+  return sentence.join(" ")
+}
