+++ src/types.d.ts
@@ -1,0 +1,4 @@
+type RunInfo = {
+  duration: number
+  numChars: number
+}
